+++ src/Dialog/index.js
@@ -1,9 +1,6 @@
 import AriaComponent from '../AriaComponent';
 import getElementPair from '../lib/getElementPair';
 import interactiveChildren from '../lib/interactiveChildren';
-import getFirstAndLastItems from '../lib/getFirstAndLastItems';
-import toArray from '../lib/toArray';
-import { tabIndexDeny, tabIndexAllow } from '../lib/rovingTabIndex';
 
 /**
  * Class to set up an interactive Dialog element.
@@ -72,7 +69,7 @@ export default class Dialog extends AriaComponent {
     const [
       firstInteractiveChild,
       lastInteractiveChild,
-    ] = getFirstAndLastItems(this.interactiveChildElements);
+    ] = this.constructor.getFirstAndLastItems(this.interactiveChildElements);
 
     // Save as instance properties.
     this.firstInteractiveChild = firstInteractiveChild;
@@ -88,7 +85,7 @@ export default class Dialog extends AriaComponent {
       this.content = Array.from(document.body.children)
         .filter((child) => ! child.contains(this.target));
     } else {
-      this.content = toArray(this.content);
+      this.content = Array.from(this.content);
     }
 
     // If no content is found.
@@ -112,7 +109,7 @@ export default class Dialog extends AriaComponent {
     this.setInteractiveChildren();
 
     // Focusable content should initially have tabindex='-1'.
-    tabIndexDeny(this.interactiveChildElements);
+    this.interactiveChildElements.forEach((item) => item.setAttribute('tabindex', '-1'));
 
     /**
      * Check if the controller is a button, but only if it doesn't already have
@@ -186,7 +183,7 @@ export default class Dialog extends AriaComponent {
       // Update target element.
       this.updateAttribute(this.target, 'aria-hidden', 'false');
 
-      tabIndexAllow(this.interactiveChildElements);
+      this.interactiveChildElements.forEach((item) => item.removeAttribute('tabindex'));
 
       document.body.addEventListener('keydown', this.handleOutsideKeydown);
 
@@ -200,7 +197,7 @@ export default class Dialog extends AriaComponent {
       this.updateAttribute(this.target, 'aria-hidden', 'true');
 
       // Focusable content should have tabindex='-1' or be removed from the DOM.
-      tabIndexDeny(this.interactiveChildElements);
+      this.interactiveChildElements.forEach((item) => item.setAttribute('tabindex', '-1'));
 
       document.body.removeEventListener('keydown', this.handleOutsideKeydown);
 
@@ -334,7 +331,7 @@ export default class Dialog extends AriaComponent {
     this.removeAttributes(this.target);
 
     // Remove tabindex attribute.
-    tabIndexAllow(this.interactiveChildElements);
+    this.interactiveChildElements.forEach((item) => item.removeAttribute('tabindex'));
 
     // Remove event listeners.
     this.controller.removeEventListener('click', this.controllerHandleClick);