+++ src/Listbox/index.js
@@ -1,7 +1,6 @@
 import Popup from '../Popup';
+import Search from './Search';
 import getElementPair from '../lib/getElementPair';
-import Search from '../lib/Search';
-import getFirstAndLastItems from '../lib/getFirstAndLastItems';
 
 /**
  * Class to set up an interactive Listbox element.
@@ -26,11 +25,6 @@ export default class ListBox extends Popup {
      * @type {string}
      */
     this[Symbol.toStringTag] = 'Listbox';
-
-    // Get the component elements.
-    const { controller, target } = getElementPair(element);
-    this.controller = controller;
-    this.target = target;
 
     // Merge options as instance properties.
     Object.assign(this, options);
@@ -75,7 +69,7 @@ export default class ListBox extends Popup {
     });
 
     // Save first and last option as properties.
-    const [firstOption, lastOption] = getFirstAndLastItems(this.options);
+    const [firstOption, lastOption] = this.constructor.getFirstAndLastItems(this.options);
     Object.assign(this, { firstOption, lastOption });
 
     /**