+++ utils/index.js
@@ -2,24 +2,9 @@
  * Select utility function exports.
  */
 import interactiveChildren from '../src/lib/interactiveChildren';
-import {
-  tabIndexAllow,
-  tabIndexDeny,
-  rovingTabIndex,
-} from '../src/lib/rovingTabIndex';
-import { getUniqueId, setUniqueId } from '../src/lib/uniqueId';
-import Search from '../src/lib/Search';
-import getFirstAndLastItems from '../src/lib/getFirstAndLastItems';
 import isInstanceOf from '../src/lib/isInstanceOf';
 
 export {
   interactiveChildren,
-  tabIndexAllow,
-  tabIndexDeny,
-  rovingTabIndex,
-  getUniqueId,
-  setUniqueId,
-  Search,
-  getFirstAndLastItems,
   isInstanceOf,
 };